add test account

+++ src/main/java/com/assistant/model/dto/ProjectDTO.java
@@ -17,7 +17,7 @@ public class ProjectDTO {
     private String project;
     private String department;
     private String nodeId;
-    private Integer avetime;
+    private Long avetime;
     private String msg;
     private State state;
